Extract utils to pkg/utils/disaggregatedset/ package
Move utility functions, types, and constants out of the controller
package into a dedicated utils package to avoid name collisions
(e.g., SetInitialReplicas exists on both the util and the manager).
